feat: 添加 JMH 性能测试
- 新增 SumTest 类，包含 4 种不同方式的求和方法
- 添加 JMH 依赖并配置 Maven 项目
- 创建简单的 web 项目结构，包括 web.xml 和 index.jsp
- 添加 .gitignore 文件，忽略常见 IDE 和系统文件

+++ src/main/java/com/github/domain/CsvRecord.java
@@ -6,7 +6,7 @@ import lombok.NoArgsConstructor;
 
 @Data
 @NoArgsConstructor
-// @JsonPropertyOrder({ "name", "surname", "shoesize", "gender" })
+// @JsonPropertyOrder({ })
 public class CsvRecord {
 
     private String id;
@@ -32,7 +32,7 @@ public class CsvRecord {
     @JsonProperty("user_id")
     private String userId;
 
-    private int age;
+    private double age;
     private String sex;
     private String local;
